requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==15.1.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dash==3.2.0
distro==1.9.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.1
flatbuffers==25.2.10
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==2.1.1
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.99.9
packaging==25.0
pandas==2.3.1
plotly==6.3.0
pluggy==1.6.0
protobuf==6.32.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyreadline3==3.5.4
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
retrying==1.4.2
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.47.3
sympy==1.14.0
text-unidecode==1.3
tokenizers==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
webrtcvad==2.0.10
Werkzeug==3.1.3
zipp==3.23.0

